feat: categories!
The ability to select a random wallpaper from any number of categories
is here! You can simply type that category you wish to use. Furthermore,
this will also create a custom path, with the folder name being the name
of the category.

This feature will likely improve over time, and the ability to specify a
custom path should be bug-free (at least from how much I've tested it).